split out helpers to utils, avoid circular import
