+++ metaseq_cli/validate.py
@@ -13,6 +13,7 @@ from itertools import chain
 import torch
 from omegaconf import DictConfig
 
+import metaseq.logging.progress_bar.utils
 from metaseq import checkpoint_utils, distributed_utils, options, utils
 from metaseq.dataclass.utils import convert_namespace_to_omegaconf
 from metaseq.logging import metrics
@@ -103,7 +104,7 @@ def main(cfg: DictConfig, override_args=None):
             num_workers=cfg.dataset.num_workers,
             data_buffer_size=cfg.dataset.data_buffer_size,
         ).next_epoch_itr(shuffle=False)
-        progress = progress_bar.progress_bar(
+        progress = metaseq.logging.progress_bar.utils.get_progress_bar(
             itr,
             log_format=cfg.common.log_format,
             log_interval=cfg.common.log_interval,
